colocate svg script w/ component
also refactor out some SVG calls to named functions so it's easier to find things relative to the UI elements
