+++ site/src/components/HomeDailyChart/generator/chart_events.json
